home move
moved home component to become a child of login component for Input() variables
