+++ web-app-frontend/src/app/login/home/home.component.html
@@ -2,5 +2,8 @@
 <div class = "container-fluid"> 
     <p>home works!</p>
     <p>welcome user: </p> 
+    <ul> 
+        <li *ngFor = "let food of foods"> {{food}}</li>
+    </ul>
 </div>
