+++ web-app-frontend/src/app/login/login.component.ts
@@ -17,9 +17,11 @@ import { ServerService } from '../services/server.service';
 })
 export class LoginComponent implements OnInit {
 
+
   //injecting formbuilder as formBuilder into component
   constructor(private formBuilder: FormBuilder, private router: Router, private backend: ServerService) { } 
 
+  user = '';
   //creating a formgroup for the form
   loginForm = this.formBuilder.group({
     //first parameter in array is the initial value
@@ -41,6 +43,7 @@ export class LoginComponent implements OnInit {
     this.backend.login(this.loginForm.value).subscribe( query => {
       if (query) { // if the login was correct
         console.log("correct login", query);
+        this.user = this.loginForm.value.user;
         this.router.navigate(['/home']); //navigate to home page
       }
       else { //if login was incorrect
